Sun Position: rename north.py to draw.py
